include: linux: Add v4l2-extensible-params.h
Import the v4l2-extensible-params.h header from the Linux
kernel sources.

The file has not been merged in mainline Linux yet but is available
at:
https://lore.kernel.org/all/20250820-extensible-parameters-validation-v4-0-30fe5a99cb1f@ideasonboard.com/

Create the include/linux/media/ directory so that header files
exported from the kernel which include this file do not need
to be adjusted when imported in libcamera.

Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>