pipeline: simple: Allow buffer counts from 1 to 16 for swISP
When using the softwareISP we allocate the output buffers ourselves,
usually from system memory. There's thus no strong reason to limit
choices for client. Situations where this might be useful include:
1. Video encoding, e.g. encoding multiple buffers in parallel.
2. Clients requesting a single buffer - e.g. in multi-stream scenarios.

Thus allow up to 16 buffers - arbitrarily decided - while keeping the
default to the previous 4.

While on it, mark the config as adjusted if we did so.

Signed-off-by: Robert Mader <robert.mader@collabora.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>