documentation: Include API docs in the sphinx documentation
This commit prepares the move of all the libcamera documetation to a
place separate from the website. This makes it easier to handle website
and docs separately.

The index is updated to mimic the structure currently available on the
website.

The build process now first creates the sphinx documentation and then
copies the doxygen documentation into the sphinx output directory.

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>