libcamera: control_serializer: Deserialize array ControlInfos
Array controls (eg. ColourCorrectionMatrix, FrameDurationLimits,
ColourGains) are serialized properly by the ControlSerializer, but are
not deserialized properly. This is because their arrayness and size are
not considered during deserialization.

Fix this by checking arrayness and size of the ControlInfo's default
value during deserialization of ControlInfoMap. Only the default value
needs to be checked as that is the only part of the ControlInfo that
will be non-scalar; min/max are defined to be scalar values.

This was not noticed before as the default value of the ControlInfo of
other array controls had been set to scalar values similar to min/max,
and ColourCorrectionMatrix was the first control to properly define a
non-scalar default value.

Other array controls that define a scalar default value need to be
fixed to define a properly sized default ControlInfo value.

Bug: https://bugs.libcamera.org/show_bug.cgi?id=285
Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>