ipa: rkisp1: Add settings for DreamChip RPPX1 ISP
The DreamChip RPPX1 ISP appears to be a superset of the Rockchip RkISP1
ISP V10. The RkISP1 interface and hardware operates at 8-bits color
depth, while the RPPX1 operates at 24-bits. For this reason the RkISP1
parameter and buffer formats are deigned around 8-bits.

The kernel driver for RPPX1 can however compensate for this and scale
down it's internal data structures to 8-bits and produce parameter
buffers (RK1E) and output statistics (RK1S) that is compatible with the
RkISP1 IPA.

The one difference is that the RPPX1 have 32 histogram buckets compared
to RkISP1 V10 16. Add a dedicated hardware revisions descriptor for the
RPPX1 device so it can reuse the existing IPA.

Signed-off-by: Niklas Söderlund <niklas.soderlund+renesas@ragnatech.se>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>