Documentation: Install API documentation within HTML directory
The libcamera documentation is made of high-level documentation and
guides, written as ReStructuredText and compiled to HTML by Sphinx, and
API reference documentation, written as comments in the code and
compiled to HTML by Doxygen.

Due to meson limitations that make it difficult to place output files in
subdirectories, the compilation process produces an html/ directory for
the Sphinx documentation, and api-html/ and internal-api-html/
directories for the Doxygen documentation. When deploying the
documentation to the libcamera.org website, the api-html and
internal-api-html/ directories are moved within html/ to make the
documentation self-contained.

The Sphinx and Doxygen documentation link to each other. The links are
generated using relative paths, based on the directory hierarchy on the
website. This makes them broken when navigating the documentation in the
build tree or in the directory where libcamera is installed.

Fix this by standardizing on the directories hierarchy of the website in
the build and install directories:

- For the build directory, we can't easily build the Doxygen
  documentation in a subdirectory of the Sphinx documentation due to
  limitations of meson. Keep the existing output directories, and
  replace the html/api-html/ and html/internal-api-html/ placeholder
  directories with symlinks to the Doxygen output directories. This is
  handled by a small custom Sphinx extension.

- For the install directory, install the Doxygen documentation to
  html/api-html/ and html/internal-api-html/. This overwrites the
  placeholders.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>