libipa: fc_queue: Add trailing underscore to private members of FrameContext
It is not immediately obvious that FCQueue accesses private members of
the FrameContext class with the help of the friend declaration. This
gets even more confusing when such a member is shadowed by a declaration
in the actual IPA FrameContext class. Improve that by accessing the
FrameContext members via references to that type. Additionally add an
underscore to the variable names like we do on other members and which
allows us to create a frame() accessor without a name clash in an
upcoming commit.

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>
Reviewed-by: Paul Elder <paul.elder@ideasonboard.com>

---

Changes in v2:
- Collected tag