libipa: agc_mean_luminance: Make startup frames and regulations speed configurable
The rkisp1 has now a good synchronization between sensor settings/isp
params and stats.  It can therefore benefit from a faster regulation. To
do that without breaking others IPAs that use AgcMeanLuminance, move
regulation speed and startup frames into member variables that can be
set at runtime.

As this requires the introduction of a protected access modifier. This
was added, so that most of the previously public functions are now also
protected, because public visibility is not needed.

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>