ipa: software_isp: AGC: do not lower gain below default gain value
At the moment when the overall image brightness is considered too high
the AGC code will lower the gain all the way down to againMin before
considering lowering the exposure.

What should happen instead is lower the gain no lower then its default
ctrl value and after that lower the exposure instead of lowering
the gain.

Otherwise there might be a heavily overexposed image (e.g. all white)
which then is made less white by a very low gain which is no good.

While at it also remove the weird limitation to only lower the gain
when exposure is set to the maximum. As long as the gain is higher
then the default gain, the gain should be lowered first.

Signed-off-by: Hans de Goede <hansg@kernel.org>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>