ipa: software_isp: AGC: Stop using delayed control for previous values
Due to a combination of not having correct control-delay information for
various sensors as well as the generic nature of the simple-pipeline +
software-ISP meaning that any pipeline delays are unknown it is impossible
to get reliable control-delay values.

Wrong control-delay values can lead to pretty bad oscilation. Since atm
it is not possible to fix the wrong control-delay values, stop using
the old sensor values reported by the delayed-controlled mechanism.

Instead remember the gain and exposures set the last time the algorithm
runs (initializing the cached values with the actual sensor values on
the first frame), combined with skipping a fixed number of frames after
changing values.

Signed-off-by: Hans de Goede <hansg@kernel.org>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>