libcamera: camera: Add `applyControls()`
Add `Camera::applyControls()` whose purpose is to apply controls as soon
as possible, without going through `Request::controls()`.

A new virtual function `PipelineHandler::applyControlsDevice()` is provided
for pipeline handler to implement fast-tracked application of controls. If
the pipeline handler does not implement that functionality, or it fails,
then a fallback mechanism is used. The controls will be saved for later,
and they will be merged into the control list of the next request sent to
the pipeline handler (`Camera::Private::waitingRequests_`).

Signed-off-by: Barnabás Pőcze <barnabas.pocze@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>