pipeline: rkisp1: Correctly handle params buffer for frame 0
The parameters for frame 0 are only active on frame 0 if the
corresponding parameter buffer is queued before STREAMON is called on
the isp. Therefore the normal mechanics of calling ipa->computeParams()
do not work, as it gets called after the first request is queued and
therefore the isp is already started. To fix that, handle the parameter
buffer the same way the initial sensor controls are handled by passing
it to the ipa->start() function and then queuing the params buffer
before starting the isp.

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>