pipeline: rkisp1: Correctly handle params buffer for frame 0
The parameters for frame 0 are only active on frame 0 if the
corresponding parameter buffer is queued before STREAMON is called on
the ISP. Therefore the normal mechanics of calling ipa->computeParams()
do not work, as it gets called after the first request is queued and
therefore the ISP is already started. To fix that, handle the parameter
buffer the same way the initial sensor controls are handled by passing
it to the ipa->start() function and then queuing the params buffer
before starting the isp.

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>

---

Changes in v3:
- Moved bufferId== 0 check out of computeParamsInternal for better
  readability
- setSensorControls is now emitted after paramsComputed
  but that must not be a problem
- Dropped rby tag due to bigger changes

Changes in v2:
- Small cleanup
- Fixed crash in raw case where a null buffer is passed on start
- Collected tag