libcamera: control_serializer: Add array info to serialized ControlValue
Array controls (eg. ColourCorrectionMatrix, FrameDurationLimits,
ColourGains) are serialized properly by the ControlSerializer, but are
not deserialized properly. This is because their arrayness and size are
not considered during deserialization.

Fix this by adding arrayness and size to the serialized form of all
ControlValues. This is achieved by adding the already-existing struct
ipa_control_value_entry to the serialized form of ControlInfoMap to
contain all the metadata associated with each ControlValue in
the min/max/def of each ControlInfo.

The issue was not noticed before as the default value of the ControlInfo
of other array controls had been set to scalar values similar to
min/max, and ColourCorrectionMatrix was the first control to properly
define a non-scalar default value.

Other array controls that define a scalar default value need to be
fixed to define a properly sized default ControlInfo value.

Bug: https://bugs.libcamera.org/show_bug.cgi?id=285
Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>