include: linux: Add v4l2-isp.h
Import the v4l2-isp.h header from the Linux kernel sources.

The file has not been merged in mainline Linux yet but is available
at v7:
https://lore.kernel.org/all/20251014-extensible-parameters-validation-v7-0-6628bed5ca98@ideasonboard.com/

Create the include/linux/media/ directory so that header files
exported from the kernel which include this file do not need
to be adjusted when imported in libcamera.

Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Tested-by: Antoine Bouyer <antoine.bouyer@nxp.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Acked-by: Kieran Bingham <kieran.bingham@ideasonboard.com>
Reviewed-by: Daniel Scally <dan.scally@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>