ipa: libipa: pwl: Add construction from initializer_list
In the tuning files it is possible to construct the PWL from a linear
list of numbers. Add a initializer_list based constructor to replicate
similar style in C++ code. An initializer_list is preferred to a Span
based constructor because the Span makes the static construction
cumbersome ( {{{ 1.0, 2.0 }}} vs { 1.0, 2.0 } ) and initialization is
the only use case at the moment.

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>