ipa: rkisp1: Update exposure limits on vblank change
The AGC algorithm implementation of the RkISP1 IPA is implemented
deriving the generic AgcMeanLuminance class, which limits the
achievable exposure time using a maxExposureTime parameter provided by
the IPA module.

The RkISP1 IPA fetches the maxExposureTime value from the IPAContext
sensor.maxExposureTime value, which is initialized at IPA init()
and configure() time, but never updated later on, effectively limiting
the achievable exposure time to the frame duration programmed at
startup time.

Whenever the frame duration is changed, the maximum exposure time
should change accordingly, but in the current implementation this
doesn't happen.

Fix this by updating the sensor.maxExposureTime value when a new set of
controls is sent to the pipeline handler. Store the current VBLANK value
in the IPAContext by replacing 'defVBlank' which is never used and use
it to detect changes in the blanking value. Whenever the blanking
changes, update the maximum exposure limit accordingly.

As reported in a comment in the code, even if the new sensor controls
will be programmed in the sensor with a delay of a few frames on the
pipeline handler side, the limits for the algorithms should be updated
immediately.

Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>