libcamera: rkisp1: Update camera::controls() on limit changes
The limits (min and max values) for the Camera controls are
registered at Camera configuration time and never updated.
Some controls, like FrameDurationLimits have a direct impact on
the limits of other controls, such as the ExposureTime and as they
can be configured by the application, this has to be taken into account.

Currently, when a user changes the frame duration, the limits for
both the ExposureTime and FrameDurationControls are not updated.

The timing at which the controls limits should be updated is also
critical: the new control values take effect once the Request they
belong to is completed.

To support this operation model introduce a new IPA function
'updateControlsLimits()' which the pipeline handler calls before
completing a Request.

Store the exposure time limits in the FrameContext (frame
duration limits were already there) and update the Camera::controls()
control info map with the limits as computed for the Request that has
just completed.

Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>