libcamera: pipeline_handler: Fix requestComplete on waiting requests on stop
The requestComplete signal is not emitted when the camera is stopped and
the request is still in the waitingRequests_ queue. This was reported in
[1]. Fix that by calling doQueueRequest() on the waiting requests after
marking them as cancelled. This ensures that the request is not queued
to the device but gets added to the queuedRequests_ list. This list is
then iterated in completeRequest() and leads to the requestComplete
signal.

[1] https://gitlab.freedesktop.org/camera/libcamera/-/issues/281

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>