libcamera: v4l2: Support string control payloads
v4l2 controls can be configured to store a string. This is stored in the
v4l2_ctrl_ext when the union within it points to a string as its
payload.

This can be read similarly to int control payloadsm instead using the
maximum length of the string, represented by info.maximum.

This pointer can then be reinterpreted as a char *.

N.B. It is possible that if the string is less than the maximum length,
the data after the string is populated with garbage. This isn't a
problem when using the string later, as the std::string will only use up
to the null character in the string.

Signed-off-by: Isaac Scott <isaac.scott@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>