libipa: Provide a Quantized type and Quantizer support
Frequently when handling data in IPA components we must convert and
store user interface values which may be floating point values, and
perform a specific operation or conversion to quantize this to a
hardware value.

This value may be to a fixed point type, or more custom code mappings,
but in either case it is important to contain both the required hardware
value, with it's effective quantized value.

Provide a new storage type 'Quantized' which can be read publicly but
must only be written through a controller class known as a Quantizer.

Quantizers can be customised and specified by the owner of the data
object and must implement conversion between floats and the underlying
hardware type.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>