ipa: camera_sensor_helper: Introduce exposureMargin()
There is a margin between the maximum achievable exposure time
and the frame length. The value is sensor-dependent and should
be used to adjust the maximum achievable exposure time whenever
the frame length changes.

Introduce in the CameraSensorHelper an exposureMargin() function
that returns a std::optional<> with the sensor-specific margin.
Use a std::optional<> to easily identify CameraSensorHelper
implementations that do not populate exposureMargin_ and warn
about possibly sub-optimal results in exposure time calculations.

This feature is already implemented in the RPi CamHelper class
hierarchy with the name of frameIntegrationDiff.

Populate the CameraSensorHelper instances with an
exposureMargin_. The value of the exposure margin comes from
the mainline driver version of each sensor, and it has been
compared with the frameIntegrationDiff_ value in the rpi
camera helpers.

The only outliner is the Vd56g3 sensor, whose exposure margin from
the driver differs from the RPi camera helper.

Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>