ipa: libipa: agc: Configure with frame duration
Pass the maximum allowed frame duration to the AGC configure()
operation.

This prepares for limiting the exposure time to the frame duration.

In order to be able to provide to the AGC the frame duration in the
IPU3 and Mali-C55 IPAs it is necessary to store the FrameDurationLimits
control in the context.

Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>