ipa: libipa: agc: Initialize a sensible frame duration
The AGC algorithm implemented by AgcMeanLuminance splits the desired
exposure between shutter time, analogue and digital gains. It tries to
maximize the shutter time up to the specified limits and then increases
the gains.

When initializing the algorithm at configure() time the IPA modules
have access to the Camera's absolute limits, which might include
a very long maximum frame duration. If such duration is used to clamp
the maximum exposure, the algorithm will maximize the shutter time,
resulting in a very low frame rate.

Try to pick a slighlty saner default of 30, 15 or 10 fps to limit the
maximum shutter time with to provide a reasonable out-of-the-box
behaviour for users. The new frame duration limit calculated by the AGC
algorithm is passed back to the IPA module that shall use it to compute
the right blankings to configure the camera with.

At the moment, only the RkISP1 IPA actually controls the frame duration,
and so it is the only IPA that needs to use the frame duration limit
computed by the AGC algorithm.

Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>