ipa: libipa: agc: Calculate frame duration in helper
The RkISP1 IPA is the only IPA that derives from AgcMeanLuminance
that populates the FrameDuration control.

The way frame duration is calculated is by using the shutter time
as returned by the Agc helpers and use that value to populate the
FrameDuration metadata and calculate the vblank.

This is however not correct as the exposure time shall be shorter than
the frame duration by a sensor-specific margin.

The AGC helpers are in charge of calculating the desired exposure
and split it between shutter time and gains. They already use the
exposure margin to limit the maximum shutter time to the maximum
frame duration. Let the helpers calculate the frame duration and
return it with shutter time and gains from calculateNewEv().

Also update the Mali C55 and IPU3 IPAs, however they do not currently
handle the frame duration and never change vblank so they don't report
the FrameDuration metadata (Mali C55) or return a fixed value (IPU3).

Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>