utils: checkstyle.py: Accept `Closes` commit trailer
`Closes: <url>` is recognized by gitlab and closes the referenced issue
automatically when a commit is made[0] to the project's default branch.
Now that issues are hosted on gitlab, it will be the preferred way to
reference the fixed issue from a commit, so add it to the checkstlye.py script.

[0]: https://docs.gitlab.com/user/project/issues/managing_issues/#default-closing-pattern

Signed-off-by: Barnabás Pőcze <barnabas.pocze@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>