ipa: rkisp1: cproc: Provide a Hue control
The RKISP1 supports a configurable Hue as part of the colour processing
unit (cproc).

This is implemented as a phase shift of the chrominance values between
-90 and +87.188 degrees according to the datasheet however the type
itself would imply that this is a range between -90 and 89.2969.

Implement the new control converting to the hardware scale accordingly
and report the applied control in the completed request metadata.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>