libcamera: controls: Strings are arrays
`ControlId::isArray()` and `ControlValue::isArray()` disagree
in the case of strings. Fix it by setting the static size of a
string to `libcamera::dynamic_extent` to denote a dynamically
sized array-like value.

One unfortunate side effect of this change is that if there were
string controls (there are none at the moment), then `cam` would
display them with an extra `Size: n` annotation.

Closes: https://gitlab.freedesktop.org/camera/libcamera/-/issues/255
Signed-off-by: Barnabás Pőcze <barnabas.pocze@ideasonboard.com>
Reviewed-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Reviewed-by: Isaac Scott <isaac.scott@ideasonboard.com>
Reviewed-by: Paul Elder <paul.elder@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>