libcamera: Add `MetadataList`
Add a dedicated `MetadataList` type, whose purpose is to store the metadata
reported by a camera for a given request. Previously, a `ControlList` was
used for this purpose. The reason for introducing a separate type is to
simplify the access to the returned metadata during the entire lifetime
of a request.

Specifically, for early metadata completion to be easily usable it should be
guaranteed that any completed metadata item can be accessed and looked up
at least until the associated requested is reused with `Request::reuse()`.

However, when a metadata item is completed early, the pipeline handler
might still work on the request in the `CameraManager`'s private thread,
therefore there is an inherent synchronization issue when an application
accesses early metadata.

Restricting the user to only access the metadata items of a not yet completed
request in the early metadata availability signal handler by ways of
documenting or enforcing it at runtime could be an option, but it is not
too convenient for the user.

The current `ControlList` implementation employs an `std::unordered_map`,
so pointers remain stable when the container is modified, so an application
could keep accessing particular metadata items outside the signal handler,
but this fact is far from obvious, and the user would still not be able
to make a copy of all metadata or do lookups based on the numeric ids or
the usual `libcamera::Control<>` objects, thus some type safety is lost.

The above also requires that each metadata item is only completed once for
a given request, but this does not appear to be serious limitation,
and in fact, this restriction is enforced by `MetadataList`.

The introduced `MetadataList` supports single writer - multiple reader
scenarios, and it can be set, looked-up, and copied in a wait-free fashion
without introducing data races or other synchronization issues. This is
achieved by requiring the possible set of metadata items to be known
(such set is stored in a `MetadataListPlan` object). Based on the this
plan, a single contiguous allocation is made to accommodate all potential
metadata items. Due to this single contiguous allocation that is not modified
during the lifetime of a `MetadataList` and atomic modifications, it is
possible to easily gaurantee thread-safe set, lookup, and copy; assuming
there is only ever a single writer.

Signed-off-by: Barnabás Pőcze <barnabas.pocze@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>