libcamera: pipeline_handler: Inject "debug" metadata
Inject all metadata controls in the "debug" namespace into
the metadata plan of each camera so that they can be used
seamlessly. Dynamically sized array-like controls have a
hard-coded size of 32 elements.

Additionally, a new type is added for inspecting properties of
a control type at runtime since that was not available previously.

Signed-off-by: Barnabás Pőcze <barnabas.pocze@ideasonboard.com>
Reviewed-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Reviewed-by: Paul Elder <paul.elder@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>