libcamera: delayed_controls: Make VBLANK priority by default
The DelayedControls class works around a limitation of the V4L2
controls API by assigning to controls that modify the limits of
other controls a 'priority' flag.

'Priority' controls are written to the device before others
to make sure the limits of dependent controls are correctly
updated.

A typical example of a priority control is VBLANK, whose value changes
the limits of the EXPOSURE control. This doesn't apply to a specific
hardware platform but to all V4L2 sensors.

It's currently up to users of delayed controls to apply to
each control the correct flag by populating a ControlParams instance.
However, this is error prone, and the RkISP1 pipeline handler
wrongly sets the VBLANK priority to false.

Instead of simply fixing the RkISP1 implementation, change how the
priorty flag is set, by making DelayedControls automatically flag
VBLANK as priority.

This limits what controls the pipelines can set as priority, but so
far only VBLANK needs that flag to be set.

Split the ControlParams type into an external one that only associates
delays to control ids and an internal one that is populated at class
creation time.

Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>