gstreamer: Plumb camera orientation configuration in libcamerasrc
Plumb the support for CameraConfiguration::orientation in libcamerasrc.
A new "orientation" property is introduced and mappings for
libcamera::Orientation <> GstVideoOrientationMethod are provided
with helpers.

If the orientation is changed after the CameraConfiguration::validate(),
a GstMessage is sent on the GstBus to report the new orientation
property value.

In case the orientation that was requested cannot be satisfied,
a 'image-orientation' tag is pushed downstream for sinks to rotate
the stream accordingly (those implementing
GSTGST_VIDEO_ORIENTATION_AUTO).

Signed-off-by: Umang Jain <uajain@igalia.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>