libcamera: debayer_cpu: Mask out unused bits from > 8bpp non packed src data
Users have been reporting invalid array access assert errors in
statsBGGR10Line0 inside the SWSTATS_ACCUMULATE_LINE_STATS() macro caused
by out of bounds accesses to the yHistogram array.

Another case of the same problem is out of bounds accesses to the various
lookup arrays used in the DebayerCpu code.

Both of these are caused by 10 bpp sparse (stored in 16 bit words) input
frames containing pixels values > 1023 leading to out of bounds array
accesses. IOW bits 15-10 of the 16 bit word are not 0 as they should be.

This can only happen if we somehow get a corrupt frame from the kernel.
The reported crashes show that this can actually happen so we will need
to harden the software ISP code to deal with this.

The cheapest (in CPU time) way to fix this is to simply mask out the high
unused bits of the words for sparse input formats directly after memcpy-ing
a line to one of the lineBuffers.

This also means that the software ISP must always use the memcpy path for
sparse 10/12 bpp input data (it should never write to the input buffer).
Add a new forceInputMemcpy_ for the setting coming from the config file and
set enableInputMemcpy_ to true if either forceInputMemcpy_ is set or
the input format is sparse 10/12 bpp.

Link: https://bugzilla.redhat.com/show_bug.cgi?id=2402746#c20
Signed-off-by: Hans de Goede <johannes.goede@oss.qualcomm.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>