treewide: Update to the C++20 standard
Switch to using the C++20 standard when compiling libcamera. This causes
a build error due to the deprecation of the following specialisation [1]

std::shared_ptr<T> atomic_load(const std::shared_ptr<T>* p)

../src/libcamera/base/log.cpp: In member function 'void libcamera::Logger::write(const libcamera::LogMessage&)':
../src/libcamera/base/log.cpp:468:61: error: 'std::shared_ptr<_Tp> std::atomic_load(const shared_ptr<_Tp>*) [with _Tp = libcamera::LogOutput]' is deprecated: use 'std::atomic<std::shared_ptr<T>>' instead [-Werror=deprecated-declarations]
  468 |         std::shared_ptr<LogOutput> output = std::atomic_load(&output_);
      |                                             ~~~~~~~~~~~~~~~~^~~~~~~~~~

This error is fixed by switching to the std::atomic<std::shared_ptr<T>>
specialisation instead. Unfortunately, this is not backward compatible
with C++17, so the fix must be included in this commit.

[1]: https://en.cppreference.com/w/cpp/memory/shared_ptr/atomic.html
Signed-off-by: Naushir Patuck <naush@raspberrypi.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>