ipa: libipa: fixedpoint: Provide a ScaledFixedPoint type
Extend the new Quantized FixedPoint capabilities with a ScaledFixedPointQTraits
which takes an existing Quantized type and scales the conversion by the
given Scale template parameter.

This can be useful to represent fixed point values which represent
a linear range with a different scale to the underlying float
or integral type.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>