ipa: mali-c55: agc: Quantise the ISP Digital Gain
The Mali-C55 ISP has a digital gain block which allows setting gain in Q5.8
format, a range of 0.0 to (very nearly) 32.0.

Convert usage to the new UQ5_8 FixedPoint Quantised type which will
support the conversion, clamping and quantisation so that the metadata
and debug prints can now report the effective gain applied instead of
the potentially inaccurate float.

As the UQ5_8 type already clamps values, remove the explicit clamping.
This removes the clamping to a minimum of 1.0 gain, so we rely on
calculateNewEv to provide a valid gain.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>