ipa: rkisp1: Store FrameDurationLimits as sensor config
The RkISP1 IPA context stores a control list ctrlMap which is used by the
AGC algorithm to initialize controls that are then added to the list of the
ones registered by the IPA module.

The FrameDurationLimits control is a bit of an outlier, as it's the only
control that is registered by the IPA in the context ctrlMap and is then
used by the AGC algorithm.

As the purpose of the context ctrlMap is to be populated by algorithms for
the IPA, do not abuse it by registering a control there in the main IPA
module and only store the control limits in the active context.

Removing FrameDurationLimits from the ctrlMap requires to store it as
a sensor configuration parameter in the IPA context and use it for
clamping the FrameDurationLimits control and to initialize the AGC
active state.

Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>