ipa: camera_sensor_helper: Introduce maxShutterTime()
There is a margin between the maximum achievable shutter time and the frame
length. The value is sensor-dependent and should be used to adjust the
maximum achievable exposure time whenever the frame length changes.

Introduce in the CameraSensorHelper the notion of exposureMargin_ as the
minimum difference in number of lines between the current frame length and
the maximum achievable shutter time and the maxShutterTime() function that
returns the maximum achievable shutter time give the current maximum frame
length.

This feature is already implemented in the RPi CamHelper class hierarchy
with the name of frameIntegrationDiff.

Populate the CameraSensorHelper instances with an exposureMargin_. The
value of the exposure margin comes from the mainline driver version of each
sensor, and it has been compared with the frameIntegrationDiff_ value in
the rpi camera helpers.

Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>