ipa: camera_sensor_helper: Introduce minFrameDuration()
After having introduced CameraSensorHelper::maxShutterTime() which
calculates the maximum shutter time which can be achieved given
a frame duration, this patch introduces the opposite operation, that
given a desired shutter time calculates what is the minimum frame
duration required to achieve it.

The intended users of this function are IPA modules that after having
calculated a new exposure time need to regulate the frame duration to
achieve it.

Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>