ipa: libipa: agc: Store sensor configuration parameters
To prepare for changing the configure() and setLimits() functions
of the ExposureModeHelper class, introduce the sensor configuration
parameters that will remain valid for a streaming session, and separate
them from the run-time configuration parameters used by the exposure
helper to split shutter time and gains.

Only introduce the type in this patch but do not make use of it yet.

This change also prepares to centralize the sensor configuration
paramters in the CameraHelper.

Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>