ipa: libipa: agc: Make Agc::configure() set limits
The AgcMeanLuminance algorithm interface has a 'configure()' and
'setLimits()' API. configure() doesn't actually do much if not
initialzing a few variables, and the Mali and IPU3 IPAs do not
even call it. Configuring the AGC requires calling setLimits() at
configure() time and at run time.

In order to prepare to differentiate between configure-time settings
of the Agc and run-time handling of dynamic parameters such as the
frame duration limits, make the configure() operation initialize
the Agc limits and provide an AgcMeanLuminance::SensorConfiguration
type for IPAs to be populated with the sensor's default values.

The new type mimics the ExposureModeHelper::SensorConfiguration on
introduced in the previous patches and both will be removed once all the
information there contained will be made available from the
CameraSensorHelper.

Update all IPAs Agc implementation deriving from AgcMeanLuminance to
populate a AgcMeanLuminance::SensorConfiguration and use it in
configure().

Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>